Fix bug with exporting collections on any grid in older Magento versions
In newer Magento versions the prepareDownloadResponse method exits
immediately, which means it's safe to place the redirectReferer() call
outside of the catch blocks. In older versions, it doesn't exit
immediately, which means the redirect always blows away the download
response headers. By moving the redirect calls into the catch blocks, it
will never be called when exceptions don't occur.