Fix incompatibilities with Magento 1.6CE+/1.11EE+
There are two places where checks are made against the old Eav_Mysql4
classes in catalog-related code. One of these is an instanceof check
that can easily be expanded to include the new Resource classes. The
other is a strict type-hint in a method signature. This type-hint
has been removed.