Add support for adding action columns to grids that don't have them
Currently it's only possible to add an edit link. This may be expanded
in the future, depending on what requirements people have. Also, you
need to manually switch on the action column by way of the custom
columns dialog.

By default, the CMS blocks and product attributes grids get this.